Move cw20-stake tests to own file, remove admin-factory profile
