+++ contracts/cw20-stake/src/lib.rs
@@ -4,4 +4,7 @@ pub mod hooks;
 pub mod msg;
 pub mod state;
 
+#[cfg(test)]
+mod tests;
+
 pub use crate::error::ContractError;
